+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes
 
 - Rename `SRC_DIR` to `WORDLISTS_DIR` (reflecting what it stands for).
 - Use also flake8 with tox.
+- Pass `options` to `get_passphrase()` instead of a bunch of single args.
 
 
 0.4 (2015-03-30)